Made Coin and Value data instances of the type family ASSET.
