+++ shelley/chain-and-ledger/executable-spec/src/Cardano/Ledger/Val.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE ConstraintKinds #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE EmptyDataDecls #-}
@@ -12,18 +13,23 @@
 {-# LANGUAGE TypeFamilies #-}
 
 -- | This module defines a generalised notion of a "value" - that is, something
--- with which we may quantify a transaction output.
+-- with which we may quantify a transaction output. Values will have Val instances
+-- and be defined as data family ASSET instances.
 module Cardano.Ledger.Val
   ( Val (..),
     scale,
     invert,
     sumVal,
-    scaledMinDeposit,
+    Asset (..),
+    ASSET,
+    Coin,
   )
 where
 
 import Data.Group (Abelian)
-import Shelley.Spec.Ledger.Coin (Coin (..))
+import Data.Kind
+
+-- ================================================================
 
 class
   ( Abelian t,
@@ -84,74 +90,11 @@ sumVal xs = foldl (<+>) mempty xs
 invert :: Val t => t -> t
 invert x = (-1 :: Integer) <×> x
 
-instance Val Coin where
-  n <×> (Coin x) = Coin $ (fromIntegral n) * x
-  coin = id
-  inject = id
-  size _ = 1
-  modifyCoin f v = f v
-  pointwise p (Coin x) (Coin y) = p x y
+-- =====================================================================
+-- Several instances of Val will be type family instances of ASSET
 
-{- The scaledMinDeposit calculation uses the minUTxOValue protocol parameter
-(passed to it as Coin mv) as a specification of "the cost of
-making a Shelley-sized UTxO entry", calculated here by "utxoEntrySizeWithoutVal + uint",
-using the constants in the "where" clause.
+data Asset = Ada | MultiAsset
 
-In the case when a UTxO entry contains coins only (and the Shelley
-UTxO entry format is used - we will extend this to be correct for other
-UTxO formats shortly), the deposit should be exactly the minUTxOValue.
-This is the "inject (coin v) == v" case.
+data family ASSET (t :: Asset) :: Type -> Type
 
-Otherwise, we calculate the per-byte deposit by multiplying the minimum deposit (which is
-for the number of Shelley UTxO-entry bytes) by the size of a Shelley UTxO entry.
-This is the "(mv * (utxoEntrySizeWithoutVal + uint))" calculation.
-
-We then calculate the total deposit required for making a UTxO entry with a Val-class
-member v by dividing "(mv * (utxoEntrySizeWithoutVal + uint))" by the
-estimated total size of the UTxO entry containing v, ie by
-"(utxoEntrySizeWithoutVal + size v)".
-
-See the formal specification for details.
-
--}
-
--- TODO : This scaling function is right for UTxO, not EUTxO
--- constants are temporary, the UTxO entry size calculation will be moved
-scaledMinDeposit :: (Val v) => v -> Coin -> Coin
-scaledMinDeposit v (Coin mv)
-  | inject (coin v) == v = Coin mv -- without non-Coin assets, scaled deposit should be exactly minUTxOValue
-  | otherwise = Coin $ fst $ quotRem (mv * (utxoEntrySizeWithoutVal + uint)) (utxoEntrySizeWithoutVal + size v) -- round down
-  where
-    -- address hash length is always same as Policy ID length
-    addrHashLen :: Integer
-    addrHashLen = 28
-
-    smallArray :: Integer
-    smallArray = 1
-
-    hashLen :: Integer
-    hashLen = 32
-
-    uint :: Integer
-    uint = 5
-
-    hashObj :: Integer
-    hashObj = 2 + hashLen
-
-    addrHeader :: Integer
-    addrHeader = 1
-
-    address :: Integer
-    address = 2 + addrHeader + 2 * addrHashLen
-
-    -- input size
-    inputSize :: Integer
-    inputSize = smallArray + uint + hashObj
-
-    -- size of output not including the Val (compute that part with vsize later)
-    outputSizeWithoutVal :: Integer
-    outputSizeWithoutVal = smallArray + address
-
-    -- size of the UTxO entry (ie the space the scaled minUTxOValue deposit pays)
-    utxoEntrySizeWithoutVal :: Integer
-    utxoEntrySizeWithoutVal = inputSize + outputSizeWithoutVal
+type Coin = ASSET 'Ada ()
