+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/Coin.hs
@@ -1,12 +1,15 @@
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Shelley.Spec.Ledger.Coin
-  ( Coin (..),
+  ( Coin,
+    ASSET (Coin),
     Core.CompactForm (..),
     DeltaCoin (..),
     word64ToCoin,
@@ -14,11 +17,13 @@ module Shelley.Spec.Ledger.Coin
     rationalToCoinViaFloor,
     addDelta,
     toDelta,
+    scaledMinDeposit,
   )
 where
 
 import Cardano.Binary (FromCBOR (..), ToCBOR (..))
 import qualified Cardano.Ledger.Core as Core
+import Cardano.Ledger.Val (ASSET, Asset (Ada), Val (..))
 import Control.DeepSeq (NFData)
 import Data.Aeson (FromJSON, ToJSON)
 import Data.Group (Abelian, Group (..))
@@ -30,7 +35,7 @@ import NoThunks.Class (NoThunks (..))
 import Quiet
 
 -- | The amount of value held by a transaction output.
-newtype Coin = Coin {unCoin :: Integer}
+newtype instance ASSET 'Ada () = Coin {unCoin :: Integer}
   deriving
     ( Eq,
       Ord,
@@ -45,6 +50,16 @@ newtype Coin = Coin {unCoin :: Integer}
   deriving (ToCBOR, FromCBOR) via Core.Compact Coin
   deriving (Semigroup, Monoid, Group, Abelian) via Sum Integer
   deriving newtype (PartialOrd)
+
+type Coin = ASSET 'Ada ()
+
+instance Val Coin where
+  n <×> (Coin x) = Coin $ (fromIntegral n) * x
+  coin = id
+  inject = id
+  size _ = 1
+  modifyCoin f v = f v
+  pointwise p (Coin x) (Coin y) = p x y
 
 newtype DeltaCoin = DeltaCoin Integer
   deriving (Eq, Ord, Generic, Enum, NoThunks, NFData, FromCBOR, ToCBOR)
@@ -85,3 +100,67 @@ instance ToCBOR (Core.CompactForm Coin) where
 
 instance FromCBOR (Core.CompactForm Coin) where
   fromCBOR = CompactCoin <$> fromCBOR
+
+-- ==========================================================
+{- The scaledMinDeposit calculation uses the minUTxOValue protocol parameter
+(passed to it as Coin mv) as a specification of "the cost of
+making a Shelley-sized UTxO entry", calculated here by "utxoEntrySizeWithoutVal + uint",
+using the constants in the "where" clause.
+
+In the case when a UTxO entry contains coins only (and the Shelley
+UTxO entry format is used - we will extend this to be correct for other
+UTxO formats shortly), the deposit should be exactly the minUTxOValue.
+This is the "inject (coin v) == v" case.
+
+Otherwise, we calculate the per-byte deposit by multiplying the minimum deposit (which is
+for the number of Shelley UTxO-entry bytes) by the size of a Shelley UTxO entry.
+This is the "(mv * (utxoEntrySizeWithoutVal + uint))" calculation.
+
+We then calculate the total deposit required for making a UTxO entry with a Val-class
+member v by dividing "(mv * (utxoEntrySizeWithoutVal + uint))" by the
+estimated total size of the UTxO entry containing v, ie by
+"(utxoEntrySizeWithoutVal + size v)".
+
+See the formal specification for details.
+-}
+
+-- TODO : This scaling function is right for UTxO, not EUTxO
+-- constants are temporary, the UTxO entry size calculation will be moved
+scaledMinDeposit :: (Val v) => v -> Coin -> Coin
+scaledMinDeposit v (Coin mv)
+  | inject (coin v) == v = Coin mv -- without non-Coin assets, scaled deposit should be exactly minUTxOValue
+  | otherwise = Coin $ fst $ quotRem (mv * (utxoEntrySizeWithoutVal + uint)) (utxoEntrySizeWithoutVal + size v) -- round down
+  where
+    -- address hash length is always same as Policy ID length
+    addrHashLen :: Integer
+    addrHashLen = 28
+
+    smallArray :: Integer
+    smallArray = 1
+
+    hashLen :: Integer
+    hashLen = 32
+
+    uint :: Integer
+    uint = 5
+
+    hashObj :: Integer
+    hashObj = 2 + hashLen
+
+    addrHeader :: Integer
+    addrHeader = 1
+
+    address :: Integer
+    address = 2 + addrHeader + 2 * addrHashLen
+
+    -- input size
+    inputSize :: Integer
+    inputSize = smallArray + uint + hashObj
+
+    -- size of output not including the Val (compute that part with vsize later)
+    outputSizeWithoutVal :: Integer
+    outputSizeWithoutVal = smallArray + address
+
+    -- size of the UTxO entry (ie the space the scaled minUTxOValue deposit pays)
+    utxoEntrySizeWithoutVal :: Integer
+    utxoEntrySizeWithoutVal = inputSize + outputSizeWithoutVal
