+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/Rewards.hs
@@ -58,7 +58,8 @@ import Shelley.Spec.Ledger.BaseTypes
     unitIntervalToRational,
   )
 import Shelley.Spec.Ledger.Coin
-  ( Coin (..),
+  ( ASSET (Coin),
+    Coin,
     coinToRational,
     rationalToCoinViaFloor,
   )
