+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/STS/Deleg.hs
@@ -35,7 +35,7 @@ import Shelley.Spec.Ledger.BaseTypes
     ShelleyBase,
     invalidKey,
   )
-import Shelley.Spec.Ledger.Coin (Coin (..))
+import Shelley.Spec.Ledger.Coin (Coin)
 import Shelley.Spec.Ledger.Credential (Credential)
 import Shelley.Spec.Ledger.Keys
   ( GenDelegPair (..),
